MAINT: restructure end-to-end tests
